% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.VLSTAR.R
\name{coef.VLSTAR}
\alias{coef.VLSTAR}
\alias{coef}
\alias{coefficients}
\title{Coefficient method for objects of class VLSTAR}
\usage{
\method{coef}{VLSTAR}(object, ...)
}
\arguments{
\item{object}{An object of class \sQuote{\code{VLSTAR}}; generated by \command{VLSTAR()}.}

\item{\dots}{Currently not used.}
}
\value{
Estimated coefficients of the VLSTAR model
}
\description{
Returns the coefficients of a VLSTAR model for objects generated by \command{VLSTAR()}
}
\examples{
mean(1:3)
}
\references{
Terasvirta T. and Yang Y. (2014), Specification, Estimation and Evaluation of Vector Smooth Transition
 Autoregressive Models with Applications. \emph{CREATES Research Paper 2014-8}
}
\author{
Andrea Bucci
}
\keyword{VLSTAR}
