\name{coef}

\alias{coef}

\alias{coefficients}

\alias{coef.VLSTAR}

\encoding{latin1}

\title{Coefficient method for objects of class VLSTAR}

\description{
  Returns the coefficients of a VLSTAR model for objects generated by
  \command{VLSTAR()}.
}

\usage{
\method{coef}{VLSTAR}(object, ...)
}

\arguments{
  \item{object}{An object of class \sQuote{\code{VLSTAR}}; generated by
    \command{VLSTAR()}} 
  \item{\dots}{Currently not used.}
}

\author{Andrea Bucci}

\references{
Terasvirta T. and Yang Y. (2014), Specification, Estimation and Evaluation of Vector Smooth Transition Autoregressive Models with Applications. \emph{CREATES Research Paper 2014-8}
}

\seealso{
  \code{\link{VLSTAR}}
}

\examples{
##
##See 'VLSTAR' examples
##
}

\keyword{VLSTAR}
