% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.VLSTAR.R
\name{logLik.VLSTAR}
\alias{logLik.VLSTAR}
\alias{logLik}
\title{Log-Likelihood method}
\usage{
\method{logLik}{VLSTAR}(object, type = c('Univariate', 'Multivariate'), \dots)
}
\arguments{
\item{object}{An object of class \sQuote{\code{VLSTAR}} obtained through \command{VLSTAR()}.}

\item{type}{Type of Log-Likelihood to be showed (univariate or multivariate).}

\item{\dots}{further arguments to be passed to and from other methods}
}
\value{
An object with class attribute \code{logLik}.
}
\description{
Returns the log-Likelihood of a VLSTAR object.
}
\details{
The log-likelihood of a VLSTAR model is defined as: \deqn{\log l(y_t|I_t;\theta)=-\frac{T\tilde{n}}{2}\ln(2\pi)-\frac{T}{2}\ln|\Omega|-\frac{1}{2}\sum_{t=1}^{T}(y_t-\tilde{G}_tB\,z_t)'\Omega^{-1}(y_t-\tilde{G}_tB\,z_t)}
}
\seealso{
\code{\link{VLSTAR}}
}
\author{
Andrea Bucci
}
