% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{loglike}
\alias{loglike}
\title{Multivariate log-likelihood}
\usage{
loglike(param, data)
}
\description{
Log-likelihood to be optimized in the ML method and used to check convergence in both methods
}
\keyword{internal}
