% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.VLSTAR.R
\name{print.VLSTAR}
\alias{print.VLSTAR}
\alias{print}
\title{Print method for objects of class VLSTAR}
\usage{
\method{print}{VLSTAR}(x, ...)
}
\arguments{
\item{x}{An object of class \sQuote{\code{VLSTAR}} obtained through \command{VLSTAR()}.}

\item{\dots}{further arguments to be passed to and from other methods}
}
\description{
\sQuote{\code{print}} methods for class \sQuote{\code{VLSTAR}}.
}
\references{
Terasvirta T. and Yang Y. (2014), Specification, Estimation and Evaluation of Vector Smooth Transition Autoregressive Models with Applications. \emph{CREATES Research Paper 2014-8}
}
\seealso{
\code{\link{VLSTAR}}
}
\author{
Andrea Bucci
}
\keyword{VLSTAR}
