\name{Sample5minutes}
\alias{Sample5minutes}
\title{
  Ten simulated prices series for 19 trading days in January 2010.
}
\description{
Ten hypothetical price series were simulated according to the factor diffusion process discussed in Barndorff-Nielsen et al.
}
\usage{
data("Sample5minutes")
}

\format{
  \code{xts} object

}

\author{
  Andrea Bucci
}


\keyword{datasets}
