% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.VLSTAR.R
\name{summary.VLSTAR}
\alias{summary.VLSTAR}
\alias{summary}
\alias{print.summary.VLSTAR}
\alias{print.summary}
\title{Summary method for objects of class VLSTAR}
\usage{
\method{summary}{VLSTAR}(object, ...)

\method{print}{summary.VLSTAR}(x, ...)
}
\arguments{
\item{object}{An object of class \sQuote{\code{VLSTAR}} obtained through \command{VLSTAR()}.}

\item{\dots}{further arguments to be passed to and from other methods}

\item{x}{A summary object of class \sQuote{\code{VLSTAR}} obtained through \command{summary()}.}
}
\description{
\sQuote{\code{summary}} methods for class \sQuote{\code{VLSTAR}}.
}
\section{Functions}{
\itemize{
\item \code{print.summary.VLSTAR}: Print of the summary
}}

\references{
Terasvirta T. and Yang Y. (2014), Specification, Estimation and Evaluation of Vector Smooth Transition Autoregressive Models with Applications. \emph{CREATES Research Paper 2014-8}
}
\seealso{
\code{\link{VLSTAR}}
}
\author{
Andrea Bucci
}
\keyword{VLSTAR}
