\name{techprices}
\alias{techprices}
\title{
  Daily closing prices of 3 tech stocks.
}
\description{
    This data set contains the series of daily prices of Google, Microsof and Amazon stocks from January 3, 2005 to June 16, 2020, gathered from Yahoo.
}
\usage{
data("techprices")
}

\format{
  An \code{xts} object with 3890 daily observations,
  ranging from from January 3, 2005 to June 16, 2020.

    \tabular{rl}{
    \code{Google} \tab daily closing prices of Google (GOOG) stock. \cr
    \code{Microsoft} \tab daily closing prices of Microsoft (MSFT) stock. \cr
    \code{Amazon} \tab daily closing stock prices of Amazon (AMZN) stock. \cr
    }

}

\author{
  Andrea Bucci
}



\keyword{datasets}
